services/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
gunicorn>=21.0.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Configuration
python-decouple>=3.8

# API Documentation
drf-yasg>=1.21.7

# AI Services
google-generativeai>=0.3.0

# Media Processing
Pillow>=10.0.0
python-magic>=0.4.27


# Production Infrastructure


# Redis Cache
redis>=5.0.0
django-redis>=5.4.0

# Celery Task Queue
celery[redis]>=5.3.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0

# RabbitMQ Event Bus
pika>=1.3.0

# Monitoring
flower>=2.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# HTTP Client (for Ollama)
requests>=2.31.0
